Split code on files
